27.2.Rendering the File List Client Side{Article References Collection Endpoint}
Well, they were - this was a total rookie mistake I made with my annotations. On the @IsGranted annotation, it should be subject=article. Try it again. Here we go - that's the error I was expecting: our favorite circular reference has been detected.

This is the exact same thing we saw a second ago when we tried to serialize a single ArticleReference