28.2.Deleting Files{The deleteFile() Service Method}
The deleteFile() Service Method

Ok: how can we delete a file? Through the magic of Flysystem of course! And the best place for that logic to live is probably UploaderHelper. We already have functions for uploading two types of files, getting the public path and reading a stream. Copy the readStream() function declaration, paste, rename it to deleteFile() and remove the return type.

We'll start the same way: by grabbing whichever filesystem we need. Next say $result = $filesystem->delete() and pass that $path. Finally, code defensively: if $result === false, throw a new exception with Error deleting "%s" and $path.