28.6.Deleting Files{Alignment Tweak}
Alignment Tweak

While we're here, let's fix this alignment issue - it's weirding me out. Down in the render() function, add a few Bootstrap classes to the download link and make the delete button smaller.

Try that. Better... but it's still just a touch off. Add vertical-align: middle to the download icon. It's subtle but... yep - the buttons are lined up now.

Next: our users are begging for another feature: the ability to rename the file after it's been uploaded.