29.1.Edit Endpoint & Deserialization{The Update API Endpoint, Deserializing JSON}
I want more fancy! Seriously, we're going to add pretty much everything we can think of to make this a sweet, flexible, sort of, file "gallery". What about allowing the user to update a file reference?

Okay, well, we're not going to allow the user to update the actual attached file, there's just no point. Want to upload a newer version of a file? Just delete the old one and upload the new one. Feature, done!

But we could allow them to change the filename. Remember: this is the original filename. And, yea, if they uploaded a file called astronaut.jpeg, it would be totally cool to let them change that to something else after. Let's do it!

The Update API Endpoint

Let's keep thinking about our ArticleReference routes as a set of nice, RESTful API endpoints. We already have an endpoint to create and delete an ArticleReference. This will be an endpoint to edit a reference... except that the only field the user will be allowed to edit will be the originalFilename.

Copy the beginning of our delete endpoint, paste, close it up and we'll call this updateArticleReference(). Keep the same URL, but change the route name to admin_article_update_reference - it should be reference, not references, let's fix that in both places - I don't think I'm referencing that route name anywhere. And instead of methods={"DELETE"}, use methods={"PUT"}.

Cool! Let's think about how we want this endpoint to work. First, our JavaScript will send a request with a JSON body that contains the data that should be updated on the ArticleReference. In this case, the data will have only one field: originalFilename.

Deserializing JSON

So far, we've been using ->json() to turn an object or multiple objects into JSON. This uses Symfony's serializer behind the scenes. Now we're going to use the serializer to do the opposite: to turn JSON back into an ArticleReference object. That's called deserialization and... it's... pretty freakin' awesome!

Let's add a few more arguments: \SerializerInterface $serializer and Request - the one from HttpFoundation - so we can read the raw JSON body.

To automagically turn the JSON into an ArticleReference object, say $serializer->deserialize(). The serializer only has these two methods: serialize() and deserialize().

This method needs the raw JSON from the request - that's $request->getContent(), what type of object to turn this into - ArticleReference::class - and the format of the data: json, because the serializer can also handle XML or any crazy format you dream up.

Finally, we can pass some options - called "context". By default, deserialize() will always create a new object... but we want it to update an existing object. To do that, pass an option called object_to_populate set to $reference.