29.2.Edit Endpoint & Deserialization
Oh, and when we've been serializing, we've been passing a groups option, which tells the serializer to put the properties from the "main" group into the JSON. We can do the same thing here: we don't want a clever user to be able to update the internal filename or the id: we need to restrict their power to changing the originalFilename.

Above $originalFilename, turn the groups value into an array and give it a second group: input.

In the controller, way back down here, set groups to input. So if any other fields or passed, they'll just be ignored.

And... yea, that's it! We do need to think about validation - but, pff, we'll handle that later - like in 2 minutes. Right now we can celebrate with $entityManager->persist($reference)... which we technically don't need because this isn't a new object, but I usually add it, and $entityManager->flush().

What should we return? Typically after you edit a resource in an API, we return that resource again. Scroll all the way up to our upload endpoint and steal the JSON logic. We could also refactor this into a private method if we wanted to avoid duplication. Back down in our method, paste, rename the variable to $reference and use 200 as the status code: we're not creating a resource in this case.

Ok, that endpoint should be good! Or at least, we're ready to hook up our JavaScript so we can find out if it explodes when we use it! That's next.