30.2.JavaScript for Editing a Reference
Previous mistake was a silly mistake. When we set the data key to the reference object, jQuery didn' send up that data as JSON, it used the standard "form submit" format. We want JSON.stringify(reference).

I think we've got it this time. Refresh, tweak the filename, click off and... no errors! Check out the network tab. Yeah 200! The response returns the updated originalFilename and, if you scroll down to the request body... cool! You can see the raw JSON that was sent up.