30.3.JavaScript for Editing a Reference{Validation}
Validation

The last thing we need to do is... add validation. I know, it's always that annoying last detail once you've got the "happy" path working perfectly. But, before modification, we could leave the filename completely blank and our system would be ok with that. Well ya know what? I am totally not ok with that!

Ultimately, our endpoint modifies the ArticleReference object and that is what we should validate. Above the originalFilename field, add @notblank() and let's also use @Length(). The length can be 255 in the database, but let's use max=100.

Then, inside our endpoint, there's no form here, but that's fine. Add the ValidatorInterface $validator argument. And right after we update the object with the serializer, add $violations = $validator->validate() and pass it the $reference object. Then if $violations->count() > 0, return $this->json($violations, 400).

We're actually not going to handle that in JavaScript - I'll leave rendering the errors up to you - you could highlight the element in red and print the error below... whatever you want.

But let's at least make sure it works. Clear out the filename, hit tab to blur and... there it is! A 400 error with our beautiful error response. To handle this in JavaScript, you'll chain a .catch() onto the end of the AJAX call and then do whatever you want.

Ok, what else can we add to our upload widget? How about the ability to reorder the list. That's next.