31.1.Reordering the Files
What else do you want to add to our file gallery widget? How about allowing them to be reordered? Yea, that isn't really related to uploading either, but a lot of people asked for it... so, let's do it!

Adding the position Field

To start, the ArticleReference entity needs a field that can store its order in the list. Find your terminal and run:
php bin/console make:entity

Update ArticleReference and add one new field position. This is an integer and make it not nullable. Cool!

Go find the property position, there it is. Make it default to 0: until the user decides to reorder stuff, setting them all to 0 is fine.