32.4.Reordering Endpoint & AJAX{Changing the Endpoint Order}
Changing the Endpoint Order

The problem were that we're not rendering the list correctly! This list loaded by making an Ajax request. In the controller... here's the endpoint: getArticleReferences(). And it get the data from $article->getArticleReferences(). The problem was that this method didn't know that it should order the reference's by position.

Open up the Article entity and, above $articleReferences, add @Orm\OrderBy({"position"="ASC"}).

Let's go check out the endpoint: I'll click to open the URL in a new tab. Woohoo! astronaut-1.jpg is third! Refresh the main page. Boom! The astronaut is right were we sorted it. Let's move it down a bit further... move the Symfony Best Practices up from the bottom and refresh. The sorting sticks. Awesome!

Next, instead of saving the uploaded files locally, let's upload them to AWS S3.